pyproject.toml
[tool.poetry]
name = "graphql-python"
version = "0.1.0"
description = "GraphQL Python server for querying PostgreSQL data"
authors = ["dongju93 <spdlqj011@gmail.com>"]
maintainers = ["dongju93 <spdlqj011@gmail.com>"]
readme = "README.md"
repository = "https://github.com/dongju93/graphql-server-py"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
strawberry-graphql = { extras = ["debug-server"], version = "^0.239.2" }
fastapi = "^0.114.0"
uvicorn = { extras = ["standard"], version = "^0.30.6" }
asyncpg = "^0.29.0"
pg-sample-schema = {path = "my_lib/pg_sample_schema-0.1.0-py3-none-any.whl"}
greenlet = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
